+++ packages/core-js-compat/src/modules-by-versions.js
@@ -1,4 +1,7 @@
 'use strict';
 module.exports = {
-  // '3.1': ['esnext.builtin.feature'],
+  3.1: [
+    'es.string.match-all',
+    'es.symbol.match-all',
+  ],
 };
