+++ packages/core-js/es/index.js
@@ -5,6 +5,7 @@ require('../modules/es.symbol.has-instance');
 require('../modules/es.symbol.is-concat-spreadable');
 require('../modules/es.symbol.iterator');
 require('../modules/es.symbol.match');
+require('../modules/es.symbol.match-all');
 require('../modules/es.symbol.replace');
 require('../modules/es.symbol.search');
 require('../modules/es.symbol.species');
@@ -75,6 +76,7 @@ require('../modules/es.string.code-point-at');
 require('../modules/es.string.ends-with');
 require('../modules/es.string.includes');
 require('../modules/es.string.match');
+require('../modules/es.string.match-all');
 require('../modules/es.string.pad-end');
 require('../modules/es.string.pad-start');
 require('../modules/es.string.repeat');
