+++ packages/core-js/es/string/virtual/match-all.js
@@ -1,0 +1,3 @@
+require('../../../modules/es.string.match-all');
+
+module.exports = require('../../../internals/entry-virtual')('String').matchAll;
