+++ packages/core-js/es/symbol/match-all.js
@@ -1,0 +1,4 @@
+require('../../modules/es.symbol.match-all');
+require('../../modules/es.string.match-all');
+
+module.exports = require('../../internals/wrapped-well-known-symbol').f('matchAll');
