+++ packages/core-js/features/string/index.js
@@ -1,5 +1,4 @@
 module.exports = require('../../es/string');
 
 require('../../modules/esnext.string.at');
-require('../../modules/esnext.string.match-all');
 require('../../modules/esnext.string.replace-all');
