+++ packages/core-js/features/symbol/match-all.js
@@ -1,0 +1,1 @@
+module.exports = require('../../es/symbol/match-all');
