+++ packages/core-js/modules/es.symbol.match-all.js
@@ -1,0 +1,2 @@
+// `Symbol.matchAll` well-known symbol
+require('../internals/define-well-known-symbol')('matchAll');
