+++ packages/core-js/proposals/string-match-all.js
@@ -1,1 +1,3 @@
-require('../modules/esnext.string.match-all');
+// TODO: Remove from `core-js@4`
+require('../modules/es.string.match-all');
+require('../modules/es.symbol.match-all');
