+++ packages/core-js/stable/string/virtual/match-all.js
