+++ packages/core-js/stage/3.js
@@ -1,5 +1,4 @@
 require('../proposals/global-this');
-require('../proposals/string-match-all');
 require('../proposals/promise-all-settled');
 
 module.exports = require('./4');
