+++ packages/core-js/stage/4.js
@@ -1,1 +1,3 @@
+require('../proposals/string-match-all');
+
 module.exports = require('../internals/path');
