+++ tests/pure/index.js
@@ -118,6 +118,7 @@ import './es.string.includes';
 import './es.string.italics';
 import './es.string.iterator';
 import './es.string.link';
+import './es.string.match-all';
 import './es.string.pad-end';
 import './es.string.pad-start';
 import './es.string.raw';
@@ -204,7 +205,6 @@ import './esnext.set.symmetric-difference';
 import './esnext.set.union';
 import './esnext.string.at';
 import './esnext.string.code-points';
-import './esnext.string.match-all';
 import './esnext.symbol.observable';
 import './esnext.symbol.pattern-match';
 import './esnext.weak-map.delete-all';
