+++ tests/tests/index.js
@@ -125,6 +125,7 @@ import './es.string.italics';
 import './es.string.iterator';
 import './es.string.link';
 import './es.string.match';
+import './es.string.match-all';
 import './es.string.pad-end';
 import './es.string.pad-start';
 import './es.string.raw';
@@ -255,7 +256,6 @@ import './esnext.set.symmetric-difference';
 import './esnext.set.union';
 import './esnext.string.at';
 import './esnext.string.code-points';
-import './esnext.string.match-all';
 import './esnext.string.replace-all';
 import './esnext.symbol.dispose';
 import './esnext.symbol.observable';
